+++ lib/Sample.h
@@ -50,7 +50,7 @@ class Sample
 
         void calculateNoriginal();                    // calculate nOriginal and nOriginalWeighted
         void setBranchAddresses(int whichCategories=1); // link the values in the tree to vars
-        float getWeight();         // get the weight for the histogram based upon the pileup weight and the MC gen weight
+        double getWeight();         // get the weight for the histogram based upon the pileup weight and the MC gen weight
 
         // get the scale factor for the MC histogram based upon the number of events, the data luminosity, and the xsec for the process 
         float getScaleFactor(float luminosity);
